Convert system test to integration
There is no need to run it in real Chrome instance
